requirements.txt
Django==5.1.1
python-dotenv==1.0.1
requests==2.32.3
waitress==3.0.2
whitenoise==6.5.0